Minor refactoring

+++ xbee-api/src/test/java/com/rapplogic/xbee/api/XBeeTest.java
@@ -15,11 +15,11 @@ class XBeeTest {
         var xbee = new XBee();
 
         // Not 0; that would be the NO_RESPONSE_FRAME_ID
-        assertThat(xbee.getNextFrameId()).isEqualTo(1);
+        assertThat(xbee.getNextFrameId()).isEqualTo((byte) 1);
 
         for (var count = 0; count < 0xFF + 1; count++) {
             assertThat(xbee.getNextFrameId()).isNotZero();
         }
-        assertThat(xbee.getNextFrameId()).isEqualTo(3);
+        assertThat(xbee.getNextFrameId()).isEqualTo((byte) 3);
     }
 }
